Conflicts:
	doc/wiki/Initial-Conditions.md
	include/Macro.h
	include/Particle.h
	src/Particle/Par_Init_ByFunction.cpp
	src/TestProblem/ELBDM/DiskHeating/Par_Init_ByFunction_DiskHeating.cpp
	src/TestProblem/ELBDM/HaloMerger/Par_Init_ByFunction_HaloMerger.cpp
	src/TestProblem/Hydro/AGORA_IsolatedGalaxy/Par_Init_ByFunction_AGORA.cpp
	src/TestProblem/Hydro/CMZ/Par_Init_ByFunction_BarredPot.cpp
	src/TestProblem/Hydro/ClusterMerger/Par_Init_ByFunction_ClusterMerger.cpp
	src/TestProblem/Hydro/ParticleEquilibriumIC/Par_Init_ByFunction_ParEqmIC.cpp
	src/TestProblem/Hydro/ParticleTest/Par_Init_ByFunction_ParticleTest.cpp
	src/TestProblem/Hydro/Plummer/Par_Init_ByFunction_Plummer.cpp
	src/TestProblem/Hydro/Zeldovich/Init_TestProb_Zeldovich.cpp

+++ include/Particle.h
@@ -51,10 +51,11 @@ void Aux_Error( const char *File, const int Line, const char *Func, const char *
 //                                          the velocity gradient is large
 //                RemoveCell              : remove particles RemoveCell-base-level-cells away from the boundary
 //                                          (for non-periodic BC only)
-//                GhostSize               : Number of ghost zones required for interpolation scheme
+//                GhostSize               : Number of ghost zones required for the interpolation scheme of massive particles
+//                GhostSizeTracer         : Number of ghost zones required for the interpolation scheme of tracer  particles
 //                NextUID                 : Next new particle UID over all MPI ranks. The UID starts from 1.
-//                AttributeFlt            : Pointer arrays to different particle float   attributes (Mass, Pos, Vel, ...)
-//                AttributeInt            : Pointer arrays to different particle integer attributes (Type)
+//                AttributeFlt            : Pointer arrays to different particle floating-point attributes (Mass, Pos, Vel, ...)
+//                AttributeInt            : Pointer arrays to different particle integer        attributes (Type)
 //                InactiveParList         : List of inactive particle IDs
 //                Mesh_Attr               : Pointer arrays to different mesh quantities mapped onto tracer particles
 //                Mesh_Attr_Num           : Number of mesh quantities mapped onto tracer particles
@@ -439,8 +440,8 @@ struct Particle_t
    //                4. Note that the global variable "AveDensity_Init" will NOT be recalculated
    //                   automatically here
    //
-   // Parameter   :  NewAttFlt : Array storing the float   attributes of new particles
-   //                NewAttInt : Array storing the integer attributes of new particles
+   // Parameter   :  NewAttFlt : Array storing the floating-point attributes of new particles
+   //                NewAttInt : Array storing the integer        attributes of new particles
    //
    // Return      :  Index of the new particle (ParID)
    //===================================================================================
@@ -465,7 +466,7 @@ struct Particle_t
                     NewAttFlt[PAR_POSX], NewAttFlt[PAR_POSY], NewAttFlt[PAR_POSZ] );
 
       if ( NewAttInt[PAR_TYPE] < (long_par)0  ||  NewAttInt[PAR_TYPE] >= (long_par)PAR_NTYPE )
-         Aux_Error( ERROR_INFO, "Incorrect particle type (%d) !!\n", (int)NewAttInt[PAR_TYPE] );
+         Aux_Error( ERROR_INFO, "Incorrect particle type (%ld) !!\n", (long)NewAttInt[PAR_TYPE] );
 #     endif
